moved topology ex/import code
